requirements.txt
# Web API framework and ASGI server.
fastapi==0.110.0
uvicorn[standard]==0.27.1
# Database ORM and PostgreSQL async driver.
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
# Security and configuration.
cryptography==42.0.5
python-dotenv==1.0.1
pydantic==2.6.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
